requirements.txt
paddlepaddle
paddlehub
visualdl
paddleseg
paddlenlp
paddleocr
numpy
